fet(55961): Exclui cobrancas e revisa bloco comentario parte 01
