+++ src/utils/ValidacoesAdicionaisFormularios.js
@@ -5,6 +5,38 @@ import moment from "moment";
 import {ASSOCIACAO_UUID} from "../services/auth.service";
 import {getPeriodoFechado} from "../services/escolas/Associacao.service";
 
+const retornaNumeroOrdinal = (index) =>{
+  let _index = index + 1;
+
+  if (_index === 10){
+    return 'Décima'
+  }else if(_index === 20){
+    return 'Vigésima'
+  }else if(_index === 30){
+    return 'Trigésima'
+  }else if(_index === 40){
+    return 'Quadragésima'
+  }else if(_index === 50){
+    return 'Quinguasésima'
+  }else if(_index === 60){
+    return 'Sextagésima'
+  }else if(_index === 70){
+    return 'Séptimagésima'
+  }else if(_index === 80){
+    return 'Octagésima'
+  }else{
+    let oridinal = _index.toOrdinal({ genero: "a"});
+    let array = oridinal.split(' ');
+    let primeira_palavra = array[0];
+    let modificada = primeira_palavra.substring(0, primeira_palavra.length - 1) + 'a';
+    if (array[1] === undefined){
+      return modificada.charAt(0).toUpperCase() + modificada.slice(1)
+    }else {
+      return modificada.charAt(0).toUpperCase() + modificada.slice(1) + " " + array[1]
+    }
+  }
+};
+
 export const checkDuplicateInObject = (propertyName, inputArray) => {
 
   var seenDuplicate = false,
